Collumn

Description

Documentation

data LayerData

Simple Datatype for storing things like connections with other Layer

data Layer

Layer Data stores a activationFunction that gets applied each time we want the output of a Layer on a current Collumn. As this activationFunction you can for example implement a SP.

data Collumn

Data for many Layers where the activationFunctions are called in order of the layers list, layer_num indicates the currently executing layers index

pushLayer :: Collumn -> Layer -> Collumn

Add a Layer to the Collumn